+++ deps/AppCore/src/win/AppWin.cpp
@@ -44,9 +44,9 @@ AppWin::~AppWin() {
 void AppWin::OnClose() {
 }
 
-void AppWin::OnResize(int width, int height) {
+void AppWin::OnResize(uint32_t width, uint32_t height) {
   if (gpu_context_)
-    gpu_context_->Resize(width, height);
+    gpu_context_->Resize((int)width, (int)height);
 }
 
 void AppWin::set_window(Ref<Window> window) {
